requirements.txt
streamlit==1.32.2
spacy==3.7.2
transformers==4.38.2
torch==2.2.1
pandas==2.0.3
pillow==10.2.0
nltk==3.8.1
beautifulsoup4==4.12.3
pysrt==1.1.2
webvtt-py==0.4.6
emoji==2.10.1
scikit-learn==1.3.2
langdetect==1.0.9
python-docx==1.1.0
rich==13.7.1
# New dependencies
sentence-transformers==2.5.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1
textblob==0.18.0
googletrans==3.1.0a0
deep-translator==1.11.4
wordcloud==1.9.3
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0
opencv-python-headless==4.9.0.80
moviepy==1.0.3
pytesseract==0.3.10
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0